Implement detection of server disconnect

+++ door-client/software/driver/src/rs485.h
@@ -24,8 +24,8 @@ class rs485 {
   public:
     rs485(uint8_t);
     byte Receive();
-    byte Send(uint8_t* data, uint8_t len = 1);
-    inline int Send(uint8_t data);
+    uint8_t Send(uint8_t* data, uint8_t len = 1);
+    inline uint8_t Send(uint8_t data);
     int Available();
     void SetDebugPort(SoftwareSerial*);
